lotto-prediction/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0
python-dateutil==2.8.2
PyJWT==2.8.0
bcrypt==4.1.2
SQLAlchemy==2.0.25
